CLeanup VMInvokable
