+++ interpreter/resources_test.go
@@ -2746,7 +2746,7 @@ func TestInterpreterDefaultDestroyEventBaseShadowing(t *testing.T) {
 		require.NoError(t, err)
 
 		// Explicitly initialize the contracts, if it's the VM.
-		if vmInvokable, ok := inter.(*test_utils.CombinedInvokable); ok {
+		if vmInvokable, ok := inter.(*test_utils.VMInvokable); ok {
 			_, err = vmInvokable.InitializeContract("base")
 			require.NoError(t, err)
 		}
