+++ test_utils/common_utils/utils.go
@@ -43,7 +43,6 @@ type Invokable interface {
 	interpreter.InvocationContext
 	interpreter.ResourceDestructionContext
 	Invoke(functionName string, arguments ...interpreter.Value) (value interpreter.Value, err error)
-	InvokeWithoutComparison(functionName string, arguments ...interpreter.Value) (value interpreter.Value, err error)
 	InvokeTransaction(arguments []interpreter.Value, signers ...interpreter.Value) error
 	GetGlobal(name string) interpreter.Value
